api/requirements.txt
CacheControl==0.11.7
falcon==1.1.0
filemagic==1.6
imdbpie==4.2.0
lockfile==0.12.2
mongoengine==0.11.0
parse-torrent-name==1.1.1
pymongo==3.4.0
python-mimeparse==1.6.0
requests==2.12.5
six==1.10.0
transmissionrpc==0.11
uwsgi-tasks==0.6.4
